+++ lib/src/dialogue/lib.dart
@@ -1,2 +1,3 @@
 export 'dialog.dart';
 export 'dialog_sheet.dart';
+export 'modal_attention.dart';